+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2025-11-26
+
+### Fixed
+- Remove all console.* calls from logger (write to file only)
+- Remove all eslint-disable comments
+- Use sentence case for all UI text
+- Simplify settings headings (remove redundant words)
+- Fix async method signatures
+
 ## [1.0.1] - 2025-11-26
 
 ### Fixed